+++ libs/trainer/mlpe/trainer/__init__.py
@@ -1,23 +1,3 @@
-import inspect
-
 from .optimizers import optimizers
 from .schedulers import schedulers
 from .trainer import train
-
-
-def _wrap_callable(callable):
-    """
-    Wrap callable so that all but its first arguments
-    can be parsed from the command line via typeo.
-    """
-
-    def func(*args, **kwargs):
-        def f(variable):
-            return callable(variable, *args, **kwargs)
-
-        return f
-
-    params = inspect.signature(callable).parameters
-    params = list(params.values())[1:]
-    func.__signature__ = inspect.Signature(params)
-    return func
