+++ libs/trainer/mlpe/trainer/optimizers.py
@@ -2,7 +2,7 @@ from typing import Iterable
 
 import torch
 
-from mlpe.trainer import _wrap_callable
+from mlpe.trainer.wrapper import _wrap_callable
 
 
 # typeo requires type hints to parse arguments from the command line.
